Fix g_warning calls to not have terminating n.
2002-03-28  Morten Welinder  <terra@diku.dk>

	* */*.c: Fix g_warning calls to not have terminating n.

+++ src/rendered-value.c
@@ -58,7 +58,7 @@ rendered_value_new (Cell *cell, MStyle const *mstyle, gboolean dynamic_width)
 	RenderedValue	*res;
 	Sheet		*sheet;
 	StyleColor	*color;
-	float		 col_width = -1.;
+	double		 col_width = -1.;
 	char *str;
 
 	g_return_val_if_fail (cell != NULL, NULL);
